splitted up main file
